refactored InlayHints to make adding new types of inlay hint easier
